% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-pcaRes.R
\docType{methods}
\name{center,pcaRes-method}
\alias{center}
\alias{center,pcaRes-method}
\title{Get the centers of the original variables}
\usage{
center(object, ...)
}
\arguments{
\item{object}{pcaRes object}

\item{...}{Not used}
}
\value{
Vector with the centers
}
\description{
Get the centers of the original variables
}
\author{
Henning Redestig
}

