\name{center}
\title{Get the centers of the original variables...}
\usage{center(object, ...)}
\description{Get the centers of the original variables}
\alias{center}
\alias{center,pcaRes-method}
\value{Vector with the centers}
\author{Henning Redestig}
\arguments{\item{object}{pcaRes object}
\item{...}{Not used}}
