% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-pcaRes.R
\docType{methods}
\name{centered,pcaRes-method}
\alias{centered}
\alias{centered,pcaRes-method}
\title{Check centering was part of the model}
\usage{
centered(object, ...)
}
\arguments{
\item{object}{pcaRes object}

\item{...}{Not used}
}
\value{
TRUE if model was centered
}
\description{
Check centering was part of the model
}
\author{
Henning Redestig
}

