\name{centered}
\title{Check centering was part of the model...}
\usage{centered(object, ...)}
\description{Check centering was part of the model}
\alias{centered}
\alias{centered,pcaRes-method}
\value{TRUE if model was centered}
\author{Henning Redestig}
\arguments{\item{object}{pcaRes object}
\item{...}{Not used}}
