% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{completeObs,nniRes-method}
\alias{completeObs}
\alias{completeObs,nniRes-method}
\alias{completeObs,pcaRes-method}
\title{Get the original data with missing values replaced with predicted
values.}
\usage{
completeObs(object, ...)
}
\arguments{
\item{object}{object to fetch complete data from}

\item{...}{Not used}
}
\value{
Completed data (matrix)
}
\description{
Get the original data with missing values replaced with predicted
values.
}
\author{
Henning Redestig
}

