\name{completeObs}
\title{Get the original data with missing values replaced with predicted...}
\usage{completeObs(object, ...)}
\description{Get the original data with missing values replaced with predicted
values.}
\alias{completeObs}
\alias{completeObs,nniRes-method}
\alias{completeObs,pcaRes-method}
\value{Completed data (matrix)}
\author{Henning Redestig}
\arguments{\item{object}{object to fetch complete data from}
\item{...}{Not used}}
