% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{cvstat,pcaRes-method}
\alias{cvstat}
\alias{cvstat,pcaRes-method}
\title{Get cross-validation statistics (e.g. \eqn{Q^2}).}
\usage{
cvstat(object, ...)
}
\arguments{
\item{object}{pcaRes object}

\item{...}{not used}
}
\value{
vector CV statistics
}
\description{
Get cross-validation statistics (e.g. \eqn{Q^2}).
}
\author{
Henning Redestig
}

