\name{cvstat}
\title{Get cross-validation statistics (e...}
\usage{cvstat(object, ...)}
\description{Get cross-validation statistics (e.g. \eqn{Q^2}).}
\value{vector CV statistics}
\alias{cvstat}
\alias{cvstat,pcaRes-method}
\author{Henning Redestig}
\arguments{\item{object}{pcaRes object}
\item{...}{not used}}
