% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forkNlpcaNet.R
\name{forkNlpcaNet}
\alias{forkNlpcaNet}
\title{Complete copy of nlpca net object}
\usage{
forkNlpcaNet(nlnet)
}
\arguments{
\item{nlnet}{a nlnet}
}
\value{
A copy of the input nlnet
}
\description{
Complete copy of nlpca net object
}
\author{
Henning Redestig
}

