% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlpca.R
\name{getHierarchicIdx}
\alias{getHierarchicIdx}
\title{Index in hiearchy}
\usage{
getHierarchicIdx(hierarchicNum)
}
\arguments{
\item{hierarchicNum}{A number}
}
\value{
...
}
\description{
Index in hiearchy
}
\author{
Henning Redestig, Matthias Scholz
}

