\name{helix}
\docType{data}
\alias{helix}
\title{A helix structured toy data set}
\description{
  simulated as data set looking like a helix
}
\usage{helix}
\format{A matrix containing 1000 observations (rows) and 
	three variables (columns).}
\source{Max Planck Institut fuer Molekulare Pflanzenphysiologie, 2005}
\keyword{datasets}
\references{
	Matthias Scholz, Fatma Kaplan, Charles L. Guy, Joachim Kopka and Joachim 	Selbig. - Non-linear PCA: a missing data approach.
	\emph{Bioinformatics 2005 21(20):3887-3895}
}

