% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{irlbaPca}
\alias{irlbaPca}
\title{Perform principal component analysis using sparse friendly
  singular value decomposition}
\usage{
irlbaPca(Matrix, nPcs = 2, varLimit = 1, verbose = interactive(),
  center = TRUE, scale = FALSE, ...)
}
\arguments{
\item{Matrix}{Un-preprocessed numerical matrix samples in rows and
variables as columns. No missing values allowed.}

\item{nPcs}{Number of components that should be extracted.}

\item{varLimit}{Optionally the ratio of variance that should be
explained. \code{nPcs} is ignored if varLimit < 1}

\item{verbose}{Verbose complaints to matrix structure}

\item{center}{Whether to center the matrix or not prior to SVD}

\item{scale}{Whether to perform uv scaling or not prior to SVD}

\item{...}{Only used for passing through arguments.}
}
\value{
A \code{pcaRes} object.
}
\description{
A wrapper function for \code{irlba::prcomp_irlba} to deliver the
result as a \code{pcaRes} method. Supplied for compatibility with
the rest of the pcaMethods package.  It is not recommended to use
this function directely but rather to use the \code{pca()} wrapper
function.
}
\examples{
data(metaboliteDataComplete)
mat <- prep(t(metaboliteDataComplete))
pc <- irlbaPca(mat, nPcs=2)
## better use pca()
pc <- pca(t(metaboliteDataComplete), method="svd", nPcs=2)
\dontshow{stopifnot(sum((fitted(pc) - mat)^2, na.rm=TRUE) < 200)}
}
\seealso{
\code{\link[irlba]{prcomp_irlba}}, \code{\link{pca}}
}
\author{
Philipp Angerer, Henning Redestig
}
\keyword{multivariate}
