% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lineSearch}
\alias{lineSearch}
\title{Line search for conjugate gradient}
\usage{
lineSearch(nlnet, dw, e0, ttGuess, trainIn, trainOut, verbose)
}
\arguments{
\item{nlnet}{The nlnet}

\item{dw}{..}

\item{e0}{..}

\item{ttGuess}{..}

\item{trainIn}{Training data}

\item{trainOut}{Fitted data}

\item{verbose}{logical, print messages}
}
\value{
...
}
\description{
Line search for conjugate gradient
}
\author{
Henning Redestig, Matthias Scholz
}

