\name{listPcaMethods}
\alias{listPcaMethods}
\title{List PCA methods}
\usage{listPcaMethods(which=c("all", "linear", "nonlinear"))}
\description{Vector with current valid PCA methods}
\value{A character vector with the current methods for doing PCA}
\author{Henning Redestig}
\arguments{\item{which}{the type of methods to get. E.g. only  get the PCA
methods based on the classical model where the fitted data is a
direct multiplication of scores and loadings.}}
