% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{loadings,ANY-method}
\alias{loadings}
\alias{loadings,ANY-method}
\title{Crude way to unmask the function with the same name from
\code{stats}}
\usage{
\S4method{loadings}{ANY}(object, ...)
}
\arguments{
\item{object}{any object}

\item{...}{not used}
}
\value{
The loadings
}
\description{
Crude way to unmask the function with the same name from
\code{stats}
}
\author{
Henning Redestig
}

