\name{loadings.pcaRes}
\alias{loadings.pcaRes}
\title{Get the loadings from a PCA model...}
\usage{loadings.pcaRes(object, ...)}
\description{Get the loadings from a PCA model}
\value{The loadings as a matrix}
\author{Henning Redestig}
\arguments{\item{object}{a pcaRes object}
\item{...}{not used}}
