\name{loadings.pcaRes}
\alias{loadings.pcaRes}
\title{Get loadings from a pcaRes object}
\usage{
  \method{loadings}{pcaRes} (object, ...)
}
\arguments{
  \item{object}{a pcaRes object}

  \item{...}{not used}
}
\value{
  The loadings as a matrix
}
\description{
  Get loadings from a pcaRes object
}
\author{
  Henning Redestig
}

