\name{loadings}
\alias{loadings}
\alias{loadings,ANY-method}
\alias{loadings,pcaRes-method}
\title{Get loadings from a pcaRes object}
\arguments{
  \item{object}{a pcaRes object}

  \item{...}{not used}

  \item{object}{any object}

  \item{...}{not used}
}
\value{
  The loadings as a matrix

  The loadings
}
\description{
  Get loadings from a pcaRes object

  Crude way to unmask the function with the same name from
  \code{stats}
}
\author{
  Henning Redestig

  Henning Redestig
}
\seealso{
  \code{\link{loadings.pcaRes}}
}

