\name{metaboliteData}
\docType{data}
\alias{metaboliteData}
\title{An incomplete metabolite data set from an Arabidopsis coldstress experiment}
\description{
	A subset of size 154 x 52 from a larger metabolite data set.
	The data contains 5\% of artificially created uniformly
	distributed misssing values.
	The data was created during an in house Arabidopsis coldstress experiment.
}
\usage{metaboliteData}
\format{A matrix containing 154 observations (rows) and 
	52 metabolites (columns).}
\source{Max Planck Institut fuer Molekulare Pflanzenphysiologie, 2005}
\keyword{datasets}
\references{
	Matthias Scholz, Fatma Kaplan, Charles L. Guy, Joachim Kopka and Joachim Selbig. 
        - Non-linear PCA: a missing data approach.
	\emph{Bioinformatics 2005 21(20):3887-3895}
}

