\name{metaboliteDataComplete}
\docType{data}
\alias{metaboliteDataComplete}
\title{A complete metabolite data set from an Arabidopsis coldstress experiment}
\description{
	A subset from a larger metabolite data set.
	This is the original, complete data set and can be used to
	compare estimation results created with the also provided
	incomplete data.
	The data was created during an in house Arabidopsis coldstress experiment.
}
\usage{metaboliteData}
\format{A matrix containing 40 observations (rows) and 
	80 metabolites (columns).}
\source{Max Planck Institut fuer Molekulare Pflanzenphysiologie, 2005}
\keyword{datasets}
\seealso{ 
  \code{\link{metaboliteData}}
}
\references{
	Matthias Scholz, Fatma Kaplan, Charles L. Guy, Joachim Kopka and Joachim 	Selbig. - Non-linear PCA: a missing data approach.
	\emph{Bioinformatics 2005 21(20):3887-3895}
}
