% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-pcaRes.R
\docType{methods}
\name{nmissing,pcaRes-method}
\alias{nmissing,pcaRes-method}
\alias{nmissing}
\alias{nmissing,nniRes-method}
\title{Missing values}
\usage{
nmissing(object, ...)
}
\arguments{
\item{object}{pcaRes object}

\item{...}{Not used}
}
\value{
Get the number of missing values
}
\description{
Missing values
}
\author{
Henning Redestig
}
