\name{nmissing}
\title{Missing values...}
\usage{nmissing(object, ...)}
\description{Missing values}
\alias{nmissing}
\alias{nmissing,pcaRes-method}
\alias{nmissing,nniRes-method}
\value{Get the number of missing values}
\author{Henning Redestig}
\arguments{\item{object}{pcaRes object}
\item{...}{Not used}}
