% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-pcaRes.R
\docType{methods}
\name{nObs,pcaRes-method}
\alias{nObs,pcaRes-method}
\alias{nObs}
\title{Get the number of observations used to build the PCA model.}
\usage{
nObs(object, ...)
}
\arguments{
\item{object}{pcaRes object}

\item{...}{Not used}
}
\value{
Number of observations
}
\description{
Get the number of observations used to build the PCA model.
}
\author{
Henning Redestig
}
