\name{nObs}
\title{Get the number of observations used to build the PCA model.}
\usage{nObs(object, ...)}
\description{Get the number of observations used to build the PCA model.}
\alias{nObs}
\alias{nObs,pcaRes-method}
\value{Number of observations}
\author{Henning Redestig}
\arguments{\item{object}{pcaRes object}
\item{...}{Not used}}
