% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-pcaRes.R
\docType{methods}
\name{nP,pcaRes-method}
\alias{nP}
\alias{nP,pcaRes-method}
\title{Get number of PCs}
\usage{
nP(object, ...)
}
\arguments{
\item{object}{pcaRes object}

\item{...}{not used}
}
\value{
Number of PCs
}
\description{
Get number of PCs
}
\author{
Henning Redestig
}

