% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{nPcs,pcaRes-method}
\alias{nPcs}
\alias{nPcs,pcaRes-method}
\title{Get number of PCs.}
\usage{
nPcs(object, ...)
}
\arguments{
\item{object}{pcaRes object}

\item{...}{not used}
}
\value{
Number of PCs
}
\description{
Get number of PCs.
}
\note{
Try to use \code{link{nP}} instead since \code{nPcs} tend to
clash with argument names.
}
\author{
Henning Redestig
}

