\name{nPcs}
\title{Get number of PCs.}
\usage{nPcs(object, ...)}
\description{Get number of PCs.}
\note{Try to use \code{link{nP}} instead since \code{nPcs} tend to
clash with argument names.}
\value{Number of PCs}
\alias{nPcs}
\alias{nPcs,pcaRes-method}
\author{Henning Redestig}
\arguments{\item{object}{pcaRes object}
\item{...}{not used}}
