% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-pcaRes.R
\docType{methods}
\name{nVar,pcaRes-method}
\alias{nVar,pcaRes-method}
\alias{nVar}
\title{Get the number of variables used to build the PCA model.}
\usage{
nVar(object, ...)
}
\arguments{
\item{object}{pcaRes object}

\item{...}{Not used}
}
\value{
Number of variables
}
\description{
Get the number of variables used to build the PCA model.
}
\author{
Henning Redestig
}
