\name{nVar}
\title{Get the number of variables used to build the PCA model.}
\usage{nVar(object, ...)}
\description{Get the number of variables used to build the PCA model.}
\alias{nVar}
\alias{nVar,pcaRes-method}
\value{Number of variables}
\author{Henning Redestig}
\arguments{\item{object}{pcaRes object}
\item{...}{}}
