\name{optiAlgCgd}
\alias{optiAlgCgd}
\title{Conjugate gradient optimization...}
\usage{optiAlgCgd(nlnet, trainIn, trainOut, verbose=FALSE)}
\description{Conjugate gradient optimization}
\value{...}
\author{Henning Redestig, Matthias Scholz}
\arguments{\item{nlnet}{The nlnet}
\item{trainIn}{Training data}
\item{trainOut}{fitted data}
\item{verbose}{logical, print messages}}
