\name{pcaMethods-deprecated}
\alias{pcaMethods-deprecated}
\title{Deprecated methods for pcaMethods}
\description{
  \describe{ \item{plotR2}{Lack of relevance for this plot
  and the fact that it can not show cross-validation based
  diagnostics in the same plot makes it redundant with the
  introduction of a dedicated \code{plot} function for
  \code{pcaRes}. The new plot only shows R2cum but the
  result is pretty much the same.}}
}
\author{
  Henning Redestig
}

