% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{pcaMethods}
\alias{pcaMethods}
\alias{pcaMethods-package}
\title{pcaMethods}
\description{
Principal Component Analysis in R
}
\details{
\tabular{ll}{
Package: \tab pcaMethods \cr
Type: \tab Package \cr
Developed since: \tab 2006 \cr
License: \tab GPL (>=3) \cr
LazyLoad: \tab yes \cr
}

Provides Bayesian PCA, Probabilistic PCA, Nipals PCA, Inverse
Non-Linear PCA and the conventional SVD PCA. A cluster  based
method for missing value estimation is included for comparison.
BPCA, PPCA and NipalsPCA may be used to perform PCA on incomplete
data as well as for accurate missing value estimation.  A set of
methods for printing and plotting the results is also provided.
All PCA methods make use of the same data structure (pcaRes) to
provide a unique interface to the PCA results. Developed at the
Max-Planck Institute for Molecular Plant Physiology, Golm,
Germany, RIKEN Plant Science Center Yokohama, Japan, and CAS-MPG
Partner Institute for Computational Biology (PICB) Shanghai,
P.R. China
}
\author{
Wolfram Stacklies, Henning Redestig
}

