\name{R2cum}
\title{Cumulative R2 is the total ratio of variance that is being...}
\description{Cumulative R2 is the total ratio of variance that is being
explained by the model}
\value{Get the cumulative R2}
\alias{R2cum}
\alias{R2cum,pcaRes-method}
\author{Henning Redestig}
\arguments{\item{object}{a \code{pcaRes} model}
\item{...}{Not used}}
