\name{repmat}
\alias{repmat}
\title{Replicate and tile an array.}
\usage{repmat(mat, M, N)}
\description{Creates a large matrix B consisting of an M-by-N tiling of copies
of A}
\value{Matrix consiting of M-by-N tiling copies of input matrix}
\author{Wolfram Stacklies}
\arguments{\item{mat}{numeric matrix}
\item{M}{number of copies in vertical direction}
\item{N}{number of copies in horizontal direction}}
