\name{scaled}
\title{Check if scaling was part of the PCA model...}
\usage{scaled(object, ...)}
\description{Check if scaling was part of the PCA model}
\alias{scaled}
\alias{scaled,pcaRes-method}
\value{TRUE if scaling was part of the PCA model}
\author{Henning Redestig}
\arguments{\item{object}{pcaRes object}
\item{...}{Not used}}
