% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-pcaRes.R
\docType{methods}
\name{scl,pcaRes-method}
\alias{scl,pcaRes-method}
\alias{scl}
\title{Get the scales (e.g. standard deviations) of the original
variables}
\usage{
scl(object, ...)
}
\arguments{
\item{object}{pcaRes object}

\item{...}{Not used}
}
\value{
Vector with the scales
}
\description{
Get the scales (e.g. standard deviations) of the original
variables
}
\seealso{
\code{\link{prep}}
}
\author{
Henning Redestig
}
