\name{scl}
\title{Get the scales (e...}
\usage{scl(object, ...)}
\description{Get the scales (e.g. standard deviations) of the original
variables}
\alias{scl}
\alias{scl,pcaRes-method}
\value{Vector with the scales}
\seealso{\code{\link{prep}}}
\author{Henning Redestig}
\arguments{\item{object}{pcaRes object}
\item{...}{Not used}}
