% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-pcaRes.R
\docType{methods}
\name{scores,pcaRes-method}
\alias{scores,pcaRes-method}
\alias{scores}
\title{Get scores from a pcaRes object}
\usage{
\S4method{scores}{pcaRes}(object, ...)
}
\arguments{
\item{object}{a pcaRes object}

\item{...}{not used}
}
\value{
The scores as a matrix
}
\description{
Get scores from a pcaRes object
}
\seealso{
\code{\link{scores.pcaRes}}
}
\author{
Henning Redestig
}
