\name{scores}
\alias{scores}
\alias{scores,pcaRes-method}
\title{Get scores from a pcaRes object}
\arguments{
  \item{object}{a pcaRes object}

  \item{...}{not used}
}
\value{
  The scores as a matrix
}
\description{
  Get scores from a pcaRes object
}
\author{
  Henning Redestig
}
\seealso{
  \code{\link{scores.pcaRes}}
}

