% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-pcaRes.R
\docType{methods}
\name{sDev,pcaRes-method}
\alias{sDev}
\alias{sDev,pcaRes-method}
\title{Get the standard deviations of the scores (indicates their
relevance)}
\usage{
sDev(object, ...)
}
\arguments{
\item{object}{pcaRes object}

\item{...}{Not used}
}
\value{
Standard devations of the scores
}
\description{
Get the standard deviations of the scores (indicates their
relevance)
}
\author{
Henning Redestig
}

