\name{sDev}
\alias{sDev}
\alias{sDev,pcaRes-method}
\title{Get the standard deviations of the scores (indicates their
relevance)}
\arguments{
  \item{object}{pcaRes object}

  \item{...}{Not used}
}
\value{
  Standard devations of the scores
}
\description{
  Get the standard deviations of the scores (indicates
  their relevance)
}
\author{
  Henning Redestig
}

