% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-pcaRes.R
\docType{methods}
\name{show-methods}
\alias{show-methods}
\alias{showPcaRes}
\alias{print,pcaRes-method}
\alias{print,nniRes-method}
\alias{show,pcaRes-method}
\alias{show,nniRes-method}
\title{Print/Show for pcaRes}
\usage{
showPcaRes(x, ...)

\S4method{print}{pcaRes}(x, ...)

\S4method{show}{pcaRes}(object)
}
\arguments{
\item{x}{a pcaRes object}

\item{...}{not used}

\item{object}{the object to print information about}
}
\value{
nothing, used for its side effect
}
\description{
Print basic information about pcaRes object
}
\author{
Henning Redestig
}
