\name{showNniRes}
\alias{showNniRes}
\title{Print a nniRes model}
\usage{
  showNniRes(x, ...)
}
\arguments{
  \item{x}{An \code{nniRes} object}

  \item{...}{Not used}
}
\value{
  Nothing, used for side-effect
}
\description{
  Print a brief description of nniRes model
}
\author{
  Henning Redestig
}

