\name{showPcaRes}
\alias{showPcaRes}
\title{Print/Show for pcaRes}
\usage{
  showPcaRes(x, ...)
}
\arguments{
  \item{x}{a pcaRes object}

  \item{...}{not used}
}
\value{
  nothing, used for its side effect
}
\description{
  Print basic information about pcaRes object
}
\author{
  Henning Redestig
}

