\name{sortFeatures}
\alias{sortFeatures}
\title{Sort the features of NLPCA object...}
\usage{sortFeatures(nlnet, trainIn, trainOut)}
\description{Sort the features of NLPCA object}
\value{...}
\author{Henning Redestig}
\arguments{\item{nlnet}{The nlnet}
\item{trainIn}{Training data in}
\item{trainOut}{Training data after it passed through the net}}
