\name{summary}
\alias{summary}
\alias{summary,pcaRes-method}
\title{Summary of PCA model}
\arguments{
  \item{object}{a pcaRes object}

  \item{...}{Not available}
}
\value{
  Nothing, used for side-effect
}
\description{
  Print a brief description of the PCA model
}
\author{
  Henning Redestig
}

