\name{vector2matrices}
\alias{vector2matrices}
\alias{vector2matrices,matrix-method}
\alias{vector2matrices,nlpcaNet-method}
\title{Tranform the vectors of weights to matrix structure}
\arguments{
  \item{object}{an nlpcaNet}

  \item{object}{an nlpcaNet}
}
\value{
  weights in matrix structure

  weights in matrix structure
}
\description{
  Tranform the vectors of weights to matrix structure

  Tranform the vectors of weights to matrix structure
}
\author{
  Henning Redestig

  Henning Redestig
}

