\name{weightsAccount}
\alias{weightsAccount}
\title{Create an object that holds the weights for nlpcaNet.}
\usage{weightsAccount(w)}
\description{Create an object that holds the weights for nlpcaNet. Holds and
sets weights in using an environment object.}
\value{A weightsAccound with \code{set} and \code{current}
functions.}
\author{Henning Redestig}
\arguments{\item{w}{\code{matrix} -- New weights}}
