\name{center}
\alias{center}
\alias{center,pcaRes-method}
\title{Get the centers of the original variables}
\usage{
  center(object, ...)
}
\arguments{
  \item{object}{pcaRes object}

  \item{...}{Not used}
}
\value{
  Vector with the centers
}
\description{
  Get the centers of the original variables
}
\author{
  Henning Redestig
}

