\name{completeObs}
\alias{completeObs}

\title{Accessor function for complete observations.}

\description{This function returns the original data (after any
  centring or scaling) but with any missing values as filled in by the
  chosen algorithm.}
\usage{completeObs(object)}
\arguments{
  \item{object}{a pcaRes object}
}

\value{
  A matrix.
}
\author{Henning Redestig}
\keyword{multivariate}
