\name{cvstat}
\alias{cvstat}
\alias{cvstat,pcaRes-method}
\title{Get cross-validation statistics (e.g. \eqn{Q^2}).}
\arguments{
  \item{object}{pcaRes object}

  \item{...}{not used}
}
\value{
  vector CV statistics
}
\description{
  Get cross-validation statistics (e.g. \eqn{Q^2}).
}
\author{
  Henning Redestig
}

