% Generated by roxygen2 (4.0.2): do not edit by hand
\name{derrorHierarchic}
\alias{derrorHierarchic}
\title{Later}
\usage{
derrorHierarchic(nlnet, trainIn, trainOut)
}
\arguments{
\item{nlnet}{the nlnet}

\item{trainIn}{training data}

\item{trainOut}{fitted data}
}
\value{
derror
}
\description{
Later
}
\author{
Henning Redestig, Matthias Scholz
}

