% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dim.pcaRes}
\alias{dim.pcaRes}
\title{Dimensions of a PCA model}
\usage{
\method{dim}{pcaRes}(x)
}
\arguments{
\item{x}{a pcaRes object}
}
\value{
Get the dimensions of this PCA model
}
\description{
Dimensions of a PCA model
}
\author{
Henning Redestig
}

