\name{errorHierarchic}
\alias{errorHierarchic}
\title{Later}
\usage{
  errorHierarchic(nlnet, trainIn, trainOut)
}
\arguments{
  \item{nlnet}{The nlnet}

  \item{trainIn}{training data}

  \item{trainOut}{fitted data}
}
\value{
  error
}
\description{
  Later
}
\author{
  Henning Redestig, Matthias Scholz
}

