% Generated by roxygen2 (4.0.2): do not edit by hand
\name{forkNlpcaNet}
\alias{forkNlpcaNet}
\title{Complete copy of nlpca net object}
\usage{
forkNlpcaNet(nlnet)
}
\arguments{
\item{nlnet}{a nlnet}
}
\value{
A copy of the input nlnet
}
\description{
Complete copy of nlpca net object
}
\author{
Henning Redestig
}

