% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getHierarchicIdx}
\alias{getHierarchicIdx}
\title{Index in hiearchy}
\usage{
getHierarchicIdx(hierarchicNum)
}
\arguments{
\item{hierarchicNum}{A number}
}
\value{
...
}
\description{
Index in hiearchy
}
\author{
Henning Redestig, Matthias Scholz
}

