% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{helix}
\alias{helix}
\title{A helix structured toy data set}
\usage{
data(helix)
}
\description{
Simulated data set looking like a helix
}
\details{
A matrix containing 1000 observations (rows) and three variables
(columns).
}
\author{
Henning Redestig
}
\references{
Matthias Scholz, Fatma Kaplan, Charles L. Guy, Joachim
Kopka and Joachim Selbig. - Non-linear PCA: a missing data
approach.  \emph{Bioinformatics 2005 21(20):3887-3895}
}
\keyword{datasets}

