% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{loadings,pcaRes-method}
\alias{loadings,pcaRes-method}
\title{Get loadings from a pcaRes object}
\usage{
\S4method{loadings}{pcaRes}(object, ...)
}
\arguments{
\item{object}{a pcaRes object}

\item{...}{not used}
}
\value{
The loadings as a matrix
}
\description{
Get loadings from a pcaRes object
}
\author{
Henning Redestig
}
\seealso{
\code{\link{loadings.pcaRes}}
}

