% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{metaboliteDataComplete}
\alias{metaboliteDataComplete}
\title{A complete metabolite data set from an Arabidopsis
coldstress experiment}
\description{
A complete subset from a larger metabolite data set.  This is the
original, complete data set and can be used to compare estimation
results created with the also provided incomplete data (called
metaboliteData).  The data was created during an in house
Arabidopsis coldstress experiment.
}
\details{
A matrix containing 154 observations (rows) and 52 metabolites
(columns).
}
\author{
Wolfram Stacklies
}
\references{
Matthias Scholz, Fatma Kaplan, Charles L. Guy, Joachim
Kopka and Joachim Selbig. - Non-linear PCA: a missing data
approach.\emph{Bioinformatics 2005 21(20):3887-3895}
}
\seealso{
\code{\link{metaboliteData}}
}
\keyword{datasets}

