% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{method,pcaRes-method}
\alias{method}
\alias{method,pcaRes-method}
\title{Get the used PCA method}
\usage{
method(object, ...)
}
\arguments{
\item{object}{pcaRes object}

\item{...}{Not used}
}
\value{
The used pca method
}
\description{
Get the used PCA method
}
\author{
Henning Redestig
}

