\name{nmissing}
\alias{nmissing}
\alias{nmissing,nniRes-method}
\alias{nmissing,pcaRes-method}
\title{Missing values}
\usage{
  nmissing(object, ...)
}
\arguments{
  \item{object}{pcaRes object}

  \item{...}{Not used}
}
\value{
  Get the number of missing values
}
\description{
  Missing values
}
\author{
  Henning Redestig
}

