% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{nObs,pcaRes-method}
\alias{nObs}
\alias{nObs,pcaRes-method}
\title{Get the number of observations used to build the PCA model.}
\usage{
nObs(object, ...)
}
\arguments{
\item{object}{pcaRes object}

\item{...}{Not used}
}
\value{
Number of observations
}
\description{
Get the number of observations used to build the PCA model.
}
\author{
Henning Redestig
}

