\name{nObs}
\alias{nObs}
\alias{nObs,pcaRes-method}
\title{Get the number of observations used to build the PCA model.}
\arguments{
  \item{object}{pcaRes object}

  \item{...}{Not used}
}
\value{
  Number of observations
}
\description{
  Get the number of observations used to build the PCA
  model.
}
\author{
  Henning Redestig
}

