% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{nVar,pcaRes-method}
\alias{nVar}
\alias{nVar,pcaRes-method}
\title{Get the number of variables used to build the PCA model.}
\usage{
nVar(object, ...)
}
\arguments{
\item{object}{pcaRes object}

\item{...}{Not used}
}
\value{
Number of variables
}
\description{
Get the number of variables used to build the PCA model.
}
\author{
Henning Redestig
}

