% Generated by roxygen2 (4.0.2): do not edit by hand
\name{optiAlgCgd}
\alias{optiAlgCgd}
\title{Conjugate gradient optimization}
\usage{
optiAlgCgd(nlnet, trainIn, trainOut, verbose = FALSE)
}
\arguments{
\item{nlnet}{The nlnet}

\item{trainIn}{Training data}

\item{trainOut}{fitted data}

\item{verbose}{logical, print messages}
}
\value{
...
}
\description{
Conjugate gradient optimization
}
\author{
Henning Redestig, Matthias Scholz
}

