\name{R2cum}
\alias{R2cum}
\alias{R2cum,pcaRes-method}
\title{Cumulative R2 is the total ratio of variance that is being
explained by the model}
\arguments{
  \item{object}{a \code{pcaRes} model}

  \item{...}{Not used}
}
\value{
  Get the cumulative R2
}
\description{
  Cumulative R2 is the total ratio of variance that is
  being explained by the model
}
\author{
  Henning Redestig
}

