\name{repmat}
\alias{repmat}
\title{Replicate and tile an array.}
\usage{
  repmat(mat, M, N)
}
\arguments{
  \item{mat}{numeric matrix}

  \item{M}{number of copies in vertical direction}

  \item{N}{number of copies in horizontal direction}
}
\value{
  Matrix consiting of M-by-N tiling copies of input matrix
}
\description{
  Creates a large matrix B consisting of an M-by-N tiling
  of copies of A
}
\author{
  Wolfram Stacklies
}

