% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{scaled,pcaRes-method}
\alias{scaled}
\alias{scaled,pcaRes-method}
\title{Check if scaling was part of the PCA model}
\usage{
scaled(object, ...)
}
\arguments{
\item{object}{pcaRes object}

\item{...}{Not used}
}
\value{
TRUE if scaling was part of the PCA model
}
\description{
Check if scaling was part of the PCA model
}
\author{
Henning Redestig
}

