\name{scaled}
\alias{scaled}
\alias{scaled,pcaRes-method}
\title{Check if scaling was part of the PCA model}
\arguments{
  \item{object}{pcaRes object}

  \item{...}{Not used}
}
\value{
  TRUE if scaling was part of the PCA model
}
\description{
  Check if scaling was part of the PCA model
}
\author{
  Henning Redestig
}

