\name{scl}
\alias{scl}
\alias{scl,pcaRes-method}
\title{Get the scales (e.g. standard deviations) of the original
variables}
\usage{
  scl(object, ...)
}
\arguments{
  \item{object}{pcaRes object}

  \item{...}{Not used}
}
\value{
  Vector with the scales
}
\description{
  Get the scales (e.g. standard deviations) of the original
  variables
}
\author{
  Henning Redestig
}
\seealso{
  \code{\link{prep}}
}

