% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{scores,pcaRes-method}
\alias{scores}
\alias{scores,pcaRes-method}
\title{Get scores from a pcaRes object}
\usage{
\S4method{scores}{pcaRes}(object, ...)
}
\arguments{
\item{object}{a pcaRes object}

\item{...}{not used}
}
\value{
The scores as a matrix
}
\description{
Get scores from a pcaRes object
}
\author{
Henning Redestig
}
\seealso{
\code{\link{scores.pcaRes}}
}

