% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scores.pcaRes}
\alias{scores.pcaRes}
\title{Get scores from a pcaRes object}
\usage{
\method{scores}{pcaRes}(object, ...)
}
\arguments{
\item{object}{a pcaRes object}

\item{...}{not used}
}
\value{
The scores as a matrix
}
\description{
Get scores from a pcaRes object
}
\author{
Henning Redestig
}

