\name{sDev}
\title{Get the standard deviations of the scores (indicates their...}
\usage{sDev(object, ...)}
\description{Get the standard deviations of the scores (indicates their
relevance)}
\alias{sDev}
\alias{sDev,pcaRes-method}
\value{Standard devations of the scores}
\author{Henning Redestig}
\arguments{\item{object}{pcaRes object}
\item{...}{Not used}}
