% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sortFeatures}
\alias{sortFeatures}
\title{Sort the features of NLPCA object}
\usage{
sortFeatures(nlnet, trainIn, trainOut)
}
\arguments{
\item{nlnet}{The nlnet}

\item{trainIn}{Training data in}

\item{trainOut}{Training data after it passed through the net}
}
\value{
...
}
\description{
Sort the features of NLPCA object
}
\author{
Henning Redestig
}

