\name{summary}
\title{Summary of PCA model}
\usage{summary(object, ...)}
\description{Print a brief description of the PCA model}
\value{Nothing, used for side-effect}
\alias{summary}
\alias{summary,pcaRes-method}
\author{Henning Redestig}
\arguments{\item{object}{a pcaRes object}
\item{...}{Not available}}
