\name{summary}
\alias{summary}
\alias{summary.pcaRes}
\alias{summary,pcaRes-method}
\title{Summary of PCA model}
\usage{
  \method{summary}{pcaRes} (object, ...)
}
\arguments{
  \item{object}{a pcaRes object}

  \item{...}{Not used}
}
\value{
  Nothing, used for side-effect
}
\description{
  Print a brief description of the PCA model
}
\author{
  Henning Redestig
}

