% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{vector2matrices,nlpcaNet-method}
\alias{vector2matrices,nlpcaNet-method}
\title{Tranform the vectors of weights to matrix structure}
\usage{
\S4method{vector2matrices}{nlpcaNet}(object)
}
\arguments{
\item{object}{an nlpcaNet}
}
\value{
weights in matrix structure
}
\description{
Tranform the vectors of weights to matrix structure
}
\author{
Henning Redestig
}

