\name{weightsAccount}
\alias{weightsAccount}
\title{Create an object that holds the weights for nlpcaNet. Holds and
sets weights in using an environment object.}
\usage{
  weightsAccount(w)
}
\arguments{
  \item{w}{\code{matrix} -- New weights}
}
\value{
  A weightsAccound with \code{set} and \code{current}
  functions.
}
\description{
  Create an object that holds the weights for nlpcaNet.
  Holds and sets weights in using an environment object.
}
\author{
  Henning Redestig
}

